requirements.txt
streamlit==1.29.0
requests==2.31.0
pandas==2.1.4
rapidfuzz==3.5.2
streamlit-searchbox==0.1.12
python-dotenv==1.0.0
aiohttp==3.9.1
streamlit-aggrid==0.3.4
plotly==5.17.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
reportlab==4.0.7
groq==0.4.1
matplotlib==3.7.2
seaborn==0.12.2
kaleido==0.2.1
Pillow==10.1.0